Add support for filtering products through a category
