+++ saleor/graphql/product/types/__init__.py
@@ -4,9 +4,12 @@ from .digital_contents import DigitalContent, DigitalContentUrl
 from .products import (
     Category,
     Collection,
+    CollectionFilterInput,
     Product,
+    ProductFilterInput,
     ProductImage,
     ProductOrder,
     ProductType,
+    ProductTypeFilterInput,
     ProductVariant,
 )
